// src/users/dto/fill-club-contract.dto.ts
export class FillClubContractDto {
  "firstName": string;
  "lastName": string;
  "gender": string;
  "fiscalCode": string;
  "birthProvince": string;
  "birthCity": string;
  "birthDate": string;
  "legalRepresentativeProvince": string;
  "legalRepresentativeZip": string;
  "legalRepresentativeCity": string;
  "legalRepresentativeAddress": string;
  "phone": string;
  "email": string;
  "docNumber": string;
  "referenceAgent": string;
  "mobile": string;
  "packCost": number;
  "currentYear": number;
  "currentDate": Date;
}